[Android] Fixed wrong casting from Sqlite-INTEGER to int (#3005)
* Fixed wrong casting

* copy fix to abi32 and changelog update

+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/SQLiteModule.java
@@ -162,7 +162,7 @@ public class SQLiteModule extends ReactContextBaseJavaModule {
       case Cursor.FIELD_TYPE_FLOAT:
         return cursor.getDouble(index);
       case Cursor.FIELD_TYPE_INTEGER:
-        return cursor.getInt(index);
+        return cursor.getLong(index);
       case Cursor.FIELD_TYPE_BLOB:
         // convert byte[] to binary string; it's good enough, because
         // WebSQL doesn't support blobs anyway
